26th Commit: Creata classe Rata, pagamento della rata, implementate le varie liste e stampe su file

+++ src/main/java/org/unict/domain/StrategyInteresseAlto.java
@@ -2,7 +2,7 @@ package org.unict.domain;
 
 public class StrategyInteresseAlto implements StrategyInteresse{
     @Override
-    public double calcolaInteresse(double importo, int numeroRate) {
-        return ((importo/numeroRate)+((importo*numeroRate*0.1)/1200));
+    public float calcolaInteresse(float importo, int numeroRate) {
+        return (float) ((importo/numeroRate)+((importo*numeroRate*0.1)/1200));
     }
 }
