+++ src/main/java/org/unict/domain/StrategyInteresseMedio.java
@@ -2,7 +2,7 @@ package org.unict.domain;
 
 public class StrategyInteresseMedio implements StrategyInteresse{
     @Override
-    public double calcolaInteresse(double importo, int numeroRate) {
-        return ((importo/numeroRate)+((importo*numeroRate*0.05)/1200));
+    public float calcolaInteresse(float importo, int numeroRate) {
+        return (float) ((importo/numeroRate)+((importo*numeroRate*0.05)/1200));
     }
 }
